+++ index.ts
@@ -7,8 +7,15 @@ export * from './src/actionFactory';
 export * from './src/helpers';
 export { debugCallbackType } from './src/debug';
 
-export let bootstrap = (defaultState: s.IState, renderCallback: () => void, debugCallback: d.debugCallbackType = undefined, subStateSeparator: string = '.') => {
-    debugCallback && d.bootstrap((m, p) => debugCallback(`fun-model -> ${m}`, p));
-    s.bootstrap(defaultState, subStateSeparator);
-    af.bootstrap(renderCallback);
+export interface IBootstrapParams {
+    debugCallback?: d.debugCallbackType;
+    withStateFreezing?: boolean | (() => boolean);
+    withExceptionHandling?: boolean | (() => boolean);
+    subStateSeparator?: string;
+}
+
+export const bootstrap = (defaultState: s.IState, onStateChanged: () => void, params: IBootstrapParams) => {
+    params.debugCallback && d.bootstrap((m, p) => params.debugCallback && params.debugCallback(`fun-model -> ${m}`, p));
+    s.bootstrap(defaultState, params.withStateFreezing, params.subStateSeparator);
+    af.bootstrap(onStateChanged, params.withExceptionHandling);
 };