+++ spec/actionFactory.spec.ts
@@ -31,16 +31,17 @@ describe('actionFactory', () => {
 
         it('changes state in array', () => {
             givenTodosStore({
-                todos: [{ done: false, name: 'First todo' }, { done: false, name: 'Second todo' }]
+                todos: [{ done: false, name: 'First todo' }, { done: false, name: 'Second todo' }],
+                nullableNumber: null
             });
 
             let testAction = af.createAction<tds.ITodo, tds.ITodoParams>(
                 {
-                    create: (params) => {
+                    create: (params: tds.ITodoParams) => {
                         return { key: `todos.${params.index}` };
                     }
                 },
-                (state, params) => { return params.todo }
+                (_state: tds.ITodo, params: tds.ITodoParams) => { return params.todo }
             );
             testAction({ index: 1, todo: { done: false, name: 'New second todo' } });
 
@@ -50,10 +51,11 @@ describe('actionFactory', () => {
 
         it('changes nested state on existing index', () => {
             givenTodosStore({
-                todos: [{ done: false, name: 'First todo' }, { done: false, name: 'Second todo' }]
+                todos: [{ done: false, name: 'First todo' }, { done: false, name: 'Second todo' }],
+                nullableNumber: null
             });
 
-            let testAction = af.createAction<boolean, number>({ create: (index) => { return { key: `todos.${index}.done` } } }, (state) => { return true; });
+            let testAction = af.createAction<boolean, number>({ create: (index) => { return { key: `todos.${index}.done` } } }, () => { return true; });
             testAction(1);
 
             expect(getTodos()[1].done).toBeTruthy();
@@ -69,6 +71,44 @@ describe('actionFactory', () => {
     });
 
     describe('createAction', () => {
+        describe('when action threw during handling and catching have not been enabled', () => {
+            let renderCallback: () => void;
+            let throwingAction: af.IParamLessAction;
+            beforeEach(() => {
+                renderCallback = jasmine.createSpy('render');
+                af.bootstrap(renderCallback, false);
+                throwingAction = af.createParamLessAction(NestedCursorTestFixture, () => {
+                    throw 'MyDummyException';
+                });
+            });
+
+            it('throws', () => {
+                expect(throwingAction).toThrow();
+            });
+        });
+
+        describe('when action threw during handling and catching have been enabled', () => {
+            let renderCallback: () => void;
+            let throwingAction: af.IParamLessAction;
+            beforeEach(() => {
+                renderCallback = jasmine.createSpy('render');
+                af.bootstrap(renderCallback, true);
+                throwingAction = af.createParamLessAction(NestedCursorTestFixture, () => {
+                    throw 'MyDummyException';
+                });
+            });
+
+            it('does not throw', () => {
+                expect(throwingAction).not.toThrow();
+            });
+
+            it('logs error', () => {
+                throwingAction();
+
+                expect(debugCallback).toHaveBeenCalledWith('Action factory has been initialized.', undefined);
+            });
+        });
+
         describe('when renderCallback has not been set', () => {
             it('does not throw if action has been only declared.', () => {
                 let testAction = af.createAction(NestedCursorTestFixture, (state: INestedState) => state);
@@ -77,7 +117,7 @@ describe('actionFactory', () => {
 
             it('throws if key does not exist', () => {
                 expect(() => {
-                    let testAction = af.createAction(NestedCursorTestFixture, (state: INestedState) => { return { state: 'new nested state' }; });
+                    let testAction = af.createParamLessAction(NestedCursorTestFixture, () => { return { state: 'new nested state' }; });
                     testAction();
                 }).toThrow('Render callback must be set before first usage through bootstrap(defaultState, () => { yourRenderCallback(); }).');
             });
@@ -94,7 +134,7 @@ describe('actionFactory', () => {
             it('does not report current state when state has not been changed.', () => {
                 givenStore(aState('nestedStateValue'));
 
-                af.createAction(NestedCursorTestFixture, (state: INestedState) => state)();
+                af.createParamLessAction(NestedCursorTestFixture, (state: INestedState) => state)();
 
                 expect(debugCallback).not.toHaveBeenCalledWith('Global state has been changed.', undefined);
             });
@@ -103,7 +143,7 @@ describe('actionFactory', () => {
                 let newState = { state: 'newValue' };
                 givenStore(aState('nestedStateValue'));
 
-                af.createAction(NestedCursorTestFixture, (state: INestedState) => newState)();
+                af.createParamLessAction(NestedCursorTestFixture, () => newState)();
 
                 expect(debugCallback).toHaveBeenCalledWith('Global state has been changed.', undefined);
             });
@@ -111,7 +151,7 @@ describe('actionFactory', () => {
             it('does not call render callback when state has not been changed', () => {
                 givenStore(aState('nestedStateValue'));
 
-                let testAction = af.createAction(NestedCursorTestFixture, (state: INestedState) => state);
+                let testAction = af.createParamLessAction(NestedCursorTestFixture, (state: INestedState) => state);
                 testAction();
 
                 expect(renderCallback).not.toHaveBeenCalled();
@@ -120,7 +160,7 @@ describe('actionFactory', () => {
             it('calls render callback when state has been changed', () => {
                 givenStore(aState('nestedStateValue'));
 
-                let testAction = af.createAction(NestedCursorTestFixture, (state: INestedState) => { return { state: 'newValue' }; })
+                let testAction = af.createParamLessAction(NestedCursorTestFixture, () => { return { state: 'newValue' }; })
                 testAction();
 
                 expect(renderCallback).toHaveBeenCalled();
@@ -129,22 +169,35 @@ describe('actionFactory', () => {
             it('calls nested actions', () => {
                 givenStore(aState('value'));
 
-                const nestedAction2 = af.createAction(NestedCursorTestFixture, (state: INestedState) => {
+                const nestedAction2 = af.createParamLessAction(NestedCursorTestFixture, (state: INestedState) => {
                     return { state: `${state.state} -> newValueFromNestedAction2` };
                 });
 
-                const nestedAction = af.createAction(NestedCursorTestFixture, (state: INestedState) => {
+                const nestedAction = af.createParamLessAction(NestedCursorTestFixture, (state: INestedState) => {
                     nestedAction2();
                     return { state: `${state.state} -> newValueFromNestedAction` };
                 });
 
-                af.createAction(NestedCursorTestFixture, (state: INestedState) => {
+                af.createParamLessAction(NestedCursorTestFixture, () => {
                     nestedAction();
                     return { state: 'newValue' };
                 })();
 
                 expect(s.getState(NestedCursorTestFixture).state)
                     .toBe('newValue -> newValueFromNestedAction -> newValueFromNestedAction2');
+            });
+
+            it('does not throw on immutability violation in no debug mode', () => {
+                d.bootstrap(undefined);
+                givenStore(aState('nestedStateValue'));
+                const testAction = af.createParamLessAction(SomeCursorTestFixture, (state: ISomeState) => {
+                    state.nested.state = state.nested.state + 'newValue';
+                    return state;
+                });
+
+                expect(() => {
+                    testAction();
+                }).not.toThrow();
             });
         });
     });
@@ -153,7 +206,7 @@ describe('actionFactory', () => {
         describe('when renderCallback has not been set', () => {
             it('throws if key does not exist', () => {
                 expect(() => {
-                    let testAction = af.createActions({
+                    let testAction = af.createParamLessActions({
                         cursor: NestedCursorTestFixture,
                         handler: (state: INestedState): INestedState => state
                     });
@@ -173,7 +226,7 @@ describe('actionFactory', () => {
             it('does not call render callback when all states have not been changed', () => {
                 givenStore(aState('nestedStateValue'));
 
-                let testAction = af.createActions(
+                let testAction = af.createParamLessActions(
                     {
                         cursor: NestedCursorTestFixture,
                         handler: (state: INestedState): INestedState => state
@@ -189,107 +242,30 @@ describe('actionFactory', () => {
             it('calls render callback when one of states has been changed', () => {
                 givenStore(aState('nestedStateValue'));
 
-                let testAction = af.createActions(
+                let testAction = af.createParamLessActions(
                     {
                         cursor: SomeCursorTestFixture,
                         handler: (state: ISomeState): ISomeState => { return { nested: state.nested, state: 'newStateValue' } }
                     },
                     {
                         cursor: NestedCursorTestFixture,
-                        handler: (state: INestedState): INestedState => { return { state: 'newNestedStateValue' }; }
+                        handler: (): INestedState => { return { state: 'newNestedStateValue' }; }
                     });
                 testAction();
 
                 expect(renderCallback).toHaveBeenCalled();
             });
-        });
-    });
 
-    describe('createAsyncAction', () => {
-        beforeEach(() => {
-            jasmine.clock().install();
-        });
-
-        describe('when renderCallback has not been set', () => {
-            it('does not throw if action has been only declared.', () => {
-                let testAction = af.createAsyncAction(NestedCursorTestFixture, (state: INestedState) => state);
-                expect(testAction).not.toBeUndefined();
-            });
-
-            it('throws if key does not exist', () => {
-                expect(() => {
-                    let testAction = af.createAsyncAction(NestedCursorTestFixture, (state: INestedState) => { return { state: 'new nested state' }; });
-                    testAction();
-                    jasmine.clock().tick(1);
-                }).toThrow('Render callback must be set before first usage through bootstrap(defaultState, () => { yourRenderCallback(); }).');
-            });
-        });
-
-        describe('when renderCallback has been set', () => {
-            let renderCallback: () => void;
-
-            beforeEach(() => {
-                renderCallback = jasmine.createSpy('render');
-                af.bootstrap(renderCallback);
-            });
-
-            it('does not report current state when state has not been changed.', () => {
+            it('uses action as simple setter when no handler defined', () => {
                 givenStore(aState('nestedStateValue'));
+                const testAction = af.createAction<string, string>(NestedStateCursorTestFixture);
 
-                let testAction = af.createAsyncAction(NestedCursorTestFixture, (state: INestedState) => state);
-                testAction();
-                jasmine.clock().tick(1);
-
-                expect(debugCallback).not.toHaveBeenCalledWith('Global state has been changed.', undefined);
-            });
-
-            it('reports state changed when debug has been enabled.', () => {
-                let newState = { state: 'newValue' };
-                givenStore(aState('nestedStateValue'));
-
-                let testAction = af.createAsyncAction(NestedCursorTestFixture, (state: INestedState) => newState);
-                testAction();
-                jasmine.clock().tick(1);
-
-                expect(debugCallback).toHaveBeenCalledWith('Global state has been changed.', undefined);
-            });
-
-            it('does not call render callback when state has not been changed', () => {
-                givenStore(aState('nestedStateValue'));
-
-                let testAction = af.createAsyncAction(NestedCursorTestFixture, (state: INestedState) => state);
-                testAction();
-                jasmine.clock().tick(1);
-
+                testAction('nestedStateValue');
                 expect(renderCallback).not.toHaveBeenCalled();
-            });
 
-            it('calls render callback when state has been changed', () => {
-                givenStore(aState('nestedStateValue'));
-
-                let testAction = af.createAsyncAction(NestedCursorTestFixture, (state: INestedState) => { return { state: 'newValue' }; });
-                testAction();
-                jasmine.clock().tick(1);
-
+                testAction('newNestedStateValue');
                 expect(renderCallback).toHaveBeenCalled();
             });
-
-            it('returns new state in resolve', (done) => {
-                givenStore(aState('nestedStateValue'));
-                let newState = { state: 'newValue' };
-
-                let testAction = af.createAsyncAction(NestedCursorTestFixture, (state: INestedState) => newState);
-                testAction()
-                    .then(r => {
-                        expect(r).toBe(newState);
-                        done();
-                    });
-                jasmine.clock().tick(1);
-            });
-        });
-
-        afterEach(() => {
-            jasmine.clock().uninstall();
         });
     });
 });
@@ -298,8 +274,8 @@ function givenStore(state: IStateTestFixture) {
     s.setState(s.rootCursor, state);
 }
 
-function resetStore() {
-    s.bootstrap({ some: { nested: { state: null } } });
+function resetStore(withFreezing: boolean = false) {
+    s.bootstrap({ some: { nested: { state: null } } }, withFreezing);
 }
 
 function aState(nestedState: string = 'aNestedState', state: string = 'aState'): IStateTestFixture {
@@ -321,6 +297,10 @@ interface INestedState extends s.IState {
 
 var NestedCursorTestFixture: s.ICursor<INestedState> = {
     key: 'some.nested'
+}
+
+var NestedStateCursorTestFixture: s.ICursor<string> = {
+    key: 'some.nested.state'
 }
 
 var SomeCursorTestFixture: s.ICursor<ISomeState> = {