+++ spec/store.spec.ts
@@ -5,11 +5,13 @@ import * as d from '../src/debug';
 describe('store', () => {
     beforeEach(() => {
         resetStore();
+
+        d.bootstrap(jasmine.createSpy('debugCallback'));
     });
 
     it('gets and sets root state if cursor key is rootStateKey', () => {
         s.bootstrap({ some: null });
-        let state = { some: 'state' };
+        const state = { some: 'state' };
 
         s.setState(s.rootCursor, state);
 
@@ -19,6 +21,44 @@ describe('store', () => {
     describe('rootCursor', () => {
         it('cursor has empty key', () => {
             expect(s.rootCursor.key).toBe('');
+        });
+    });
+
+    describe('isExistingCursor', () => {
+        describe('for Object', () => {
+            beforeEach(() => {
+                s.bootstrap({ some: { nested: { state: 'value' } } });
+            });
+
+            it('returns true when value exists in state through cursor', () => {
+                const exists = s.isExistingCursor({ key: 'some.nested.state' });
+
+                expect(exists).toBeTruthy();
+            });
+
+            it('returns false when value does not exist in state through cursor', () => {
+                const exists = s.isExistingCursor({ key: 'some.nested.notExistingState' });
+
+                expect(exists).toBeFalsy();
+            });
+        });
+        describe('for Array', () => {
+            beforeEach(() => {
+                s.bootstrap({ some: { nested: { arrayState: ['value1', 'value2'] } } });
+            });
+
+            it('returns true when value exists in state through cursor', () => {
+                debugger;
+                const exists = s.isExistingCursor({ key: 'some.nested.arrayState.0' });
+
+                expect(exists).toBeTruthy();
+            });
+
+            it('returns false when value does not exist in state through cursor', () => {
+                const exists = s.isExistingCursor({ key: 'some.nested.arrayState.5' });
+
+                expect(exists).toBeFalsy();
+            });
         });
     });
 
@@ -38,7 +78,7 @@ describe('store', () => {
             it('returns nested state by cursor', () => {
                 givenStore({ some: { nested: { state: 'value' } } });
 
-                let state = s.getState({ key: 'some.nested.state' });
+                const state = s.getState({ key: 'some.nested.state' });
 
                 expect(state).toBe('value');
             });
@@ -46,7 +86,7 @@ describe('store', () => {
             it('returns nested state by cursor when current value of sub state is empty string', () => {
                 givenStore({ some: { nested: { state: '' } } });
 
-                let state = s.getState({ key: 'some.nested.state' });
+                const state = s.getState({ key: 'some.nested.state' });
 
                 expect(state).toBe('');
             });
@@ -54,11 +94,6 @@ describe('store', () => {
             it('throws if key does not exist', () => {
                 expect(() => s.getState<s.IState>({ key: 'not.existing.key' }))
                     .toThrow('State for cursor key (not.existing.key) does not exist.');
-            });
-
-            it('throws if cursor key is null', () => {
-                expect(() => s.getState<s.IState>({ key: null }))
-                    .toThrow('Cursor key cannot be null.');
             });
         });
 
@@ -68,17 +103,23 @@ describe('store', () => {
             });
 
             it('returns nested state in the array on specified index', () => {
-                givenTodoStore({ todos: [{ done: false, name: 'First Todo' }, { done: false, name: 'Second Todo' }] });
+                givenTodoStore({ 
+                    todos: [{ done: false, name: 'First Todo' }, { done: false, name: 'Second Todo' }] ,
+                    nullableNumber: null
+                });
 
-                let state = s.getState<tds.ITodosState>({ key: 'todos.1.name' });
+                const state = s.getState<tds.ITodosState>({ key: 'todos.1.name' });
 
                 expect(state).toBe('Second Todo');
             });
 
             it('returns full array when is as last key', () => {
-                givenTodoStore({ todos: [{ done: false, name: 'First Todo' }] });
+                givenTodoStore({ 
+                    todos: [{ done: false, name: 'First Todo' }],
+                    nullableNumber: null
+                });
 
-                let state = s.getState<tds.ITodosState>(tds.todosCursor);
+                const state = s.getState<tds.ITodo[]>(tds.todosCursor);
 
                 expect(state[0].done).toBeFalsy();
             });
@@ -98,27 +139,35 @@ describe('store', () => {
         });
 
         describe('with booting', () => {
-            let rootCursorTestFixture: s.ICursor<IStateTestFixture> = {
+            const rootCursorTestFixture: s.ICursor<IStateTestFixture> = {
                 key: ''
             }
 
             beforeEach(() => {
-                s.bootstrap({ key: null });
+                s.bootstrap({ key: null }, true);
             });
 
-            it('throws if cursor key is null', () => {
-                expect(() => s.setState({ key: null }, {}))
-                    .toThrow('Cursor key cannot be null.');
+            it('creates empty object if cursor has not existing key', () => {
+                const cursor = { key: 'invalid' };
+                s.setState(cursor, {});
+
+                expect(s.getState(s.rootCursor)).toEqual({ key: null, 'invalid': {} });
             });
 
-            it('throws if cursor has not existing key', () => {
-                expect(() => s.setState({ key: 'invalid' }, {}))
-                    .toThrow('State for cursor key (invalid) does not exist.');
+            it('creates neseted empty objects if cursor has not existing key', () => {
+                const cursor = { key: 'not.existing.key' };
+                s.setState(cursor, {});
+
+                expect(s.getState(s.rootCursor)).toEqual({ key: null, 'not': { 'existing': { key: {} } } });
             });
 
-            it('throws if cursor has nested not existing key', () => {
-                expect(() => s.setState({ key: 'not.existing.key' }, {}))
-                    .toThrow('State for cursor key (not.existing.key) does not exist.');
+            it('freezes state', () => {
+                givenStore({ some: { nested: { state: 'value' } } });
+
+                let state = { nested: { state: 'newValue' } };
+                s.setState({ key: 'some' }, state);
+
+                expect(Object.isFrozen(state)).toBeTruthy();
             });
 
             it('sets nested state by cursor', () => {
@@ -138,7 +187,7 @@ describe('store', () => {
             });
 
             it('sets new instance of root state', () => {
-                let initState = { some: { nested: { state: 'value' } } };
+                const initState = { some: { nested: { state: 'value' } } };
                 givenStore(initState);
 
                 s.setState({ key: 'some' }, { nested: { state: 'newValue' } });
@@ -147,7 +196,7 @@ describe('store', () => {
             });
 
             it('logs new global state when debuging has been enabled', () => {
-                let debugCallback = jasmine.createSpy('debugCallback');
+                const debugCallback = jasmine.createSpy('debugCallback');
                 d.bootstrap(debugCallback);
                 givenStore({ some: { nested: { state: 'value' } } });
                 s.setState({ key: 'some' }, { nested: { state: 'newValue' } });
@@ -156,12 +205,12 @@ describe('store', () => {
             });
 
             it('sets new instances of nodes which is in the path to root state', () => {
-                let initState = { some: { nested: { state: 'value' } } };
+                const initState = { some: { nested: { state: 'value' } } };
                 givenStore(initState);
 
                 s.setState({ key: 'some.nested.state' }, 'newValue');
 
-                let newState = s.getState(rootCursorTestFixture);
+                const newState = s.getState(rootCursorTestFixture);
                 expect(newState).not.toBe(initState);
                 expect(newState.some).not.toBe(initState.some);
                 expect(newState.some.nested).not.toBe(initState.some.nested);
@@ -174,7 +223,10 @@ describe('store', () => {
             });
 
             it('sets nested state into array on specified index', () => {
-                givenTodoStore({ todos: [{ done: false, name: 'First Todo' }, { done: false, name: 'Second Todo' }] });
+                givenTodoStore({ 
+                    todos: [{ done: false, name: 'First Todo' }, { done: false, name: 'Second Todo' }],
+                    nullableNumber: null
+                });
 
                 s.setState({ key: 'todos.1.name' }, 'New Todo Name');
 
@@ -182,7 +234,10 @@ describe('store', () => {
             });
 
             it('sets new state into array on specified index', () => {
-                givenTodoStore({ todos: [{ done: false, name: 'First Todo' }] });
+                givenTodoStore({ 
+                    todos: [{ done: false, name: 'First Todo' }],
+                    nullableNumber: null
+                });
 
                 s.setState({ key: 'todos.0' }, { done: false, name: 'Second Todo' });
 
@@ -190,8 +245,11 @@ describe('store', () => {
             });
 
             it('sets new instance of array when nested item has been changed', () => {
-                let storedTodos = [{ done: false, name: 'First Todo' }, { done: false, name: 'Second Todo' }];
-                givenTodoStore({ todos: storedTodos });
+                const storedTodos = [{ done: false, name: 'First Todo' }, { done: false, name: 'Second Todo' }];
+                givenTodoStore({ 
+                    todos: storedTodos,
+                    nullableNumber: null
+                });
 
                 s.setState({ key: 'todos.1.name' }, 'New Todo Name');
 
@@ -199,7 +257,10 @@ describe('store', () => {
             });
 
             it('sets full array', () => {
-                givenTodoStore({ todos: [{ done: false, name: 'First Todo' }] });
+                givenTodoStore({ 
+                    todos: [{ done: false, name: 'First Todo' }],
+                    nullableNumber: null
+                });
 
                 s.setState(tds.todosCursor, [{ done: false, name: 'Second Todo' }, { done: false, name: 'Third Todo' }]);
 
@@ -212,14 +273,44 @@ describe('store', () => {
                 s.setState(s.rootCursor, state);
             }
         });
+
+        describe('with booting and nullableNumber cursor', () => {
+            beforeEach(() => {
+                s.bootstrap(tds.default());
+            });
+
+            it('read null from state', () => {
+                const value = s.getState(tds.nullableNumberCursor);
+
+                expect(value).toBeNull();
+            });
+
+            it('write value to state', () => {
+                s.setState(tds.nullableNumberCursor, 10);
+
+                expect(s.getState(tds.nullableNumberCursor)).toBe(10);
+            });
+
+            it('write null to state', () => {
+                s.setState(s.rootCursor, <tds.ITodosState>{ 
+                    todos: [{ done: false, name: 'First Todo' }],
+                    nullableNumber: 10
+                });
+
+                s.setState(tds.nullableNumberCursor, null);
+
+                expect(s.getState(tds.nullableNumberCursor)).toBeNull();
+            });
+        });
     });
 
     function givenStore(state: IStateTestFixture) {
         s.setState(s.rootCursor, state);
     }
 
-    function resetStore() {
-        s.bootstrap(null);
+
+    function resetStore(withFreezing: boolean = false) {
+        s.bootstrap(null, withFreezing);
     }
 });