+++ src/store.ts
@@ -4,56 +4,96 @@ import * as d from './debug';
 export interface IState {
 }
 
-export interface ICursor<TState extends IState> {
+export interface ICursor<TState extends IState | null> {
     key: string;
+    _?: TState;
 }
 
 export interface ICursorFactory<TState, TParms> {
     create(data: TParms): ICursor<TState>;
 }
 
-let state: IState = null;
+let state: IState | null = null;
 let stateSeparator = '.';
-let rootStateKey = '';
+let freezing: boolean | (() => boolean) = false;
+const rootStateKey = '';
 
-export let rootCursor: ICursor<IState> = {
+export const rootCursor: ICursor<IState> = {
     key: rootStateKey
 };
 
-export let bootstrap = (defaultState: IState, subStateSeparator: string = '.') => {
+export const bootstrap = (defaultState: IState | null, withStateFreezing: boolean | (() => boolean) = false, subStateSeparator: string = '.') => {
     stateSeparator = subStateSeparator;
     state = defaultState;
+    freezing = withStateFreezing;
 };
 
-export let getState = <TState extends IState>(cursor: ICursor<TState>): TState => {
-    let getInnerState = (innerState: IState, path: string[]): IState => {
+export const isExistingCursor = <TState extends IState | null>(cursor: ICursor<TState>): boolean => {
+    const hasExistingInnerStateInArray = (innerState: IState[], path: string[]): boolean => {
+        const index = Number(path.shift());
+        return index < innerState.length
+            ? hasExistingInnerState(innerState[index], path)
+            : false;
+    }
+    const hasExistingInnerState = (innerState: IState, path: string[]): boolean => {
+        if (path.length === 0)
+            return true;
+        const subPath = path.shift();
+        if (!subPath)
+            return true;
+        if ((<any>innerState)[subPath] === undefined)
+            return false;
+        const prop = (<any>innerState)[subPath];
+        return Array.isArray(prop)
+            ? hasExistingInnerStateInArray(prop, path)
+            : hasExistingInnerState(prop, path);
+    };
+
+    if (!isSetDefaultState(state) || !isValidCursorKey(cursor))
+        throw 'Invalid operation.';
+
+    return cursor.key === rootStateKey
+        ? true
+        : hasExistingInnerState(state, cursor.key.split(stateSeparator));
+}
+
+export const getState = <TState extends IState | null>(cursor: ICursor<TState>): TState => {
+    const getInnerState = (innerState: IState, path: string[]): IState => {
         if (path.length === 0)
             return innerState;
-        let subPath = path.shift();
+        const subPath = path.shift();
+        if (!subPath)
+            return innerState;
         checkSubstate(innerState, subPath, cursor.key);
-        let prop = innerState[subPath];
+        const prop = (<any>innerState)[subPath];
         return Array.isArray(prop) && path.length > 0
             ? getInnerState(prop[Number(path.shift())], path)
             : getInnerState(prop, path);
     };
-    checkDefaultStateAndCursor(cursor);
+
+    if (!isSetDefaultState(state) || !isValidCursorKey(cursor))
+        throw 'Invalid operation.';
+
     return <TState>(cursor.key === rootStateKey
         ? state
         : getInnerState(state, cursor.key.split(stateSeparator)));
 };
 
-export let setState = <TState extends IState>(cursor: ICursor<TState>, updatedState: TState) => {
-    let setInnerState = <TInnerState extends IState>(innerState: TInnerState, path: string[]): TInnerState => {
+export const setState = <TState extends IState | null>(cursor: ICursor<TState>, updatedState: TState) => {
+    const setInnerState = <TInnerState extends IState | null>(innerState: TInnerState, path: string[]): TInnerState => {
         if (path.length === 0)
             return <any>updatedState;
-        let subPath = path.shift();
-        checkSubstate(innerState, subPath, cursor.key);
-        let prop = innerState[subPath];
-        let newSubState = null;
+        const subPath = path.shift();
+        if (!subPath)
+            return <any>updatedState;
+
+        createSubstate(innerState, subPath);
+        const prop = (<any>innerState)[subPath];
+        let newSubState: Object | Array<IState> | null = null;
         if (Array.isArray(prop) && path.length > 0) {
             let index = Number(path.shift());
-            prop[index] = setInnerState(prop[index], path);
             newSubState = [...prop];
+            (<any>newSubState)[index] = setInnerState((<any>newSubState)[index], path);
         }
         else
             newSubState = setInnerState(prop, path);
@@ -61,29 +101,41 @@ export let setState = <TState extends IState>(cursor: ICursor<TState>, updatedSt
         if (newSubState === prop)
             return innerState;
 
-        let newState = h.shallowCopy(innerState);
-        newState[subPath] = newSubState;
+        const newState = h.shallowCopy(innerState);
+        (<any>newState)[subPath] = newSubState;
         return newState;
     };
 
-    checkDefaultStateAndCursor(cursor);
+    if (!isSetDefaultState(state) || !isValidCursorKey(cursor))
+        throw 'Invalid operation.';
 
     state =
         cursor.key === rootStateKey
             ? updatedState
             : setInnerState(state, cursor.key.split(stateSeparator));
+    if (h.isFunction(freezing) ? freezing() : freezing)
+        h.deepFreeze(state);
     d.log('Current state:', state);
 };
 
-function checkSubstate(s: IState, subPath: string, cursorKey: string) {
-    if (s[subPath] === undefined)
+function checkSubstate<TState extends IState | null>(s: TState, subPath: string, cursorKey: string) {
+    if ((<any>s)[subPath] === undefined)
         throw `State for cursor key (${cursorKey}) does not exist.`;
 }
 
-function checkDefaultStateAndCursor<TState extends IState>(cursor: ICursor<TState>) {
+function createSubstate<TState extends IState | null>(s: TState, subPath: string) {
+    if ((<any>s)[subPath] === undefined)
+        (<any>s)[subPath] = {};
+}
+
+function isSetDefaultState<T>(state: T | null): state is T {
     if (state === null)
         throw 'Default state must be set before first usage through bootstrap(defaultState, () => { yourRenderCallback(); }).';
+    return true;
+}
 
+function isValidCursorKey<TState extends IState | null>(cursor: ICursor<TState>): boolean {
     if (cursor.key === null)
         throw 'Cursor key cannot be null.';
+    return true;
 }